agents/style_assistant/requirements.txt
mlflow==3.6.0
databricks-langchain
databricks-vectorsearch
databricks-agents
